refactor: move price methods to the LedgerEngine class
